# artifacts/api-server/src/routes/ai.ts
import { Router, type IRouter } from "express";
import { db, notesTable } from "@workspace/db";
import { eq, and } from "drizzle-orm";
import {
  GenerateSummaryParams,
  GenerateQuestionsParams,
  GenerateFlashcardsParams,
  GenerateQuizParams,
} from "@workspace/api-zod";
import { requireAuth, type AuthRequest } from "../middlewares/auth";

const router: IRouter = Router();

// Dummy AI responses — ready to be replaced with Gemini API calls

router.post("/notes/:noteId/summary", requireAuth, async (req: AuthRequest, res): Promise<void> => {
  const params = GenerateSummaryParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [note] = await db
    .select()
    .from(notesTable)
    .where(and(eq(notesTable.id, params.data.noteId), eq(notesTable.userId, req.userId!)));

  if (!note) {
    res.status(404).json({ error: "Note not found" });
    return;
  }

  res.json({
    noteId: note.id,
    summary: `This note covers the key concepts from "${note.title}". The material explores foundational ideas and their practical applications. Students should focus on understanding the core principles and how they relate to real-world scenarios. The content is structured to build from basic definitions to more complex applications, making it suitable for exam preparation.`,
    keyPoints: [
      "Understand the core definitions and terminology introduced in this topic",
      "Be able to explain the main concepts in your own words",
      "Know how to apply these principles to solve practical problems",
      "Recognize common patterns and exceptions to the general rules",
      "Connect this material to related topics covered earlier in the course",
    ],
    generatedAt: new Date().toISOString(),
  });
});

router.post("/notes/:noteId/questions", requireAuth, async (req: AuthRequest, res): Promise<void> => {
  const params = GenerateQuestionsParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [note] = await db
    .select()
    .from(notesTable)
    .where(and(eq(notesTable.id, params.data.noteId), eq(notesTable.userId, req.userId!)));

  if (!note) {
    res.status(404).json({ error: "Note not found" });
    return;
  }

  res.json([
    {
      id: 1,
      question: `What are the fundamental principles underlying the concepts discussed in "${note.title}"?`,
      answer: "The fundamental principles include understanding the core definitions, recognizing patterns, and applying theoretical knowledge to practical scenarios. Students must be able to articulate these principles clearly and demonstrate their application through examples.",
      difficulty: "medium",
    },
    {
      id: 2,
      question: "How do the main concepts in this topic relate to one another, and what are the key distinctions?",
      answer: "The concepts are interconnected through shared foundational ideas while maintaining distinct characteristics. Understanding these relationships helps build a comprehensive mental model of the subject.",
      difficulty: "hard",
    },
    {
      id: 3,
      question: "Define the key terminology introduced in this material and provide examples for each.",
      answer: "Each term has a precise definition that distinguishes it from related concepts. Providing concrete examples demonstrates a deeper understanding beyond rote memorization.",
      difficulty: "easy",
    },
    {
      id: 4,
      question: "What are the practical applications of the theories and methods described in this note?",
      answer: "Practical applications span multiple domains and include problem-solving techniques, analytical methods, and real-world implementations that demonstrate the utility of theoretical knowledge.",
      difficulty: "medium",
    },
    {
      id: 5,
      question: "What common misconceptions exist about this topic, and how can they be corrected?",
      answer: "Common misconceptions arise from oversimplification or incomplete understanding. Correcting them requires careful analysis of definitions and testing assumptions against known examples and counterexamples.",
      difficulty: "hard",
    },
  ]);
});

router.post("/notes/:noteId/flashcards", requireAuth, async (req: AuthRequest, res): Promise<void> => {
  const params = GenerateFlashcardsParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [note] = await db
    .select()
    .from(notesTable)
    .where(and(eq(notesTable.id, params.data.noteId), eq(notesTable.userId, req.userId!)));

  if (!note) {
    res.status(404).json({ error: "Note not found" });
    return;
  }

  res.json([
    {
      id: 1,
      front: "What is the primary focus of this topic?",
      back: "The primary focus is on understanding core concepts and their relationships, enabling students to apply knowledge effectively in both theoretical and practical contexts.",
      topic: "Introduction",
    },
    {
      id: 2,
      front: "What method is used to analyze the main subject?",
      back: "A systematic approach is used, breaking down complex ideas into components, analyzing each part, and then synthesizing a comprehensive understanding of the whole.",
      topic: "Methodology",
    },
    {
      id: 3,
      front: "What are the three key components of this concept?",
      back: "1. The foundational definition and scope\n2. The underlying mechanisms or processes\n3. The practical applications and real-world implications",
      topic: "Core Concepts",
    },
    {
      id: 4,
      front: "How does this material connect to prerequisite knowledge?",
      back: "This material builds directly on prerequisite concepts by extending foundational ideas, introducing more nuanced variations, and demonstrating how earlier principles apply in more complex situations.",
      topic: "Connections",
    },
    {
      id: 5,
      front: "What is the most important formula or rule in this topic?",
      back: "The key rule states that understanding context is essential before applying any formula or method. The formula itself must be interpreted within the specific domain and constraints of the problem.",
      topic: "Key Rules",
    },
    {
      id: 6,
      front: "What distinguishes an expert understanding from a beginner's?",
      back: "Expert understanding involves recognizing patterns quickly, knowing when exceptions apply, connecting ideas across different contexts, and being able to explain concepts clearly to others.",
      topic: "Mastery",
    },
  ]);
});

router.post("/notes/:noteId/quiz", requireAuth, async (req: AuthRequest, res): Promise<void> => {
  const params = GenerateQuizParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [note] = await db
    .select()
    .from(notesTable)
    .where(and(eq(notesTable.id, params.data.noteId), eq(notesTable.userId, req.userId!)));

  if (!note) {
    res.status(404).json({ error: "Note not found" });
    return;
  }

  res.json([
    {
      id: 1,
      question: `Which statement best describes the main theme of "${note.title}"?`,
      options: [
        "A surface-level overview with no practical applications",
        "A comprehensive study of core concepts with practical applications",
        "A historical analysis with no modern relevance",
        "An introduction to unrelated advanced topics",
      ],
      correctIndex: 1,
      explanation: "The material provides a comprehensive study that balances theoretical foundations with practical applications, making it directly relevant to exam preparation and real-world use.",
    },
    {
      id: 2,
      question: "When applying the concepts from this note, what should be your first step?",
      options: [
        "Jump directly to the solution using memorized formulas",
        "Identify the type of problem and relevant concepts",
        "Skip the theoretical background and focus on examples",
        "Rely solely on intuition without systematic analysis",
      ],
      correctIndex: 1,
      explanation: "Identifying the type of problem and mapping it to relevant concepts is the systematic first step that ensures you apply the correct approach and avoid common mistakes.",
    },
    {
      id: 3,
      question: "Which of the following is NOT a characteristic of a strong understanding of this material?",
      options: [
        "Ability to explain concepts in simple terms",
        "Recognition of common patterns and exceptions",
        "Memorizing every detail without understanding context",
        "Connecting ideas to related topics",
      ],
      correctIndex: 2,
      explanation: "Memorizing without understanding context leads to brittle knowledge that fails when problems are presented in unfamiliar ways. Deep understanding is marked by flexibility and the ability to explain and apply concepts.",
    },
    {
      id: 4,
      question: "What is the most effective strategy for mastering the key points in this note?",
      options: [
        "Read the material once right before the exam",
        "Active recall through practice problems and self-testing",
        "Passive re-reading of the same content repeatedly",
        "Focusing exclusively on definitions while ignoring examples",
      ],
      correctIndex: 1,
      explanation: "Active recall through practice problems and self-testing is proven to be far more effective than passive reading. It strengthens memory pathways and reveals gaps in understanding.",
    },
    {
      id: 5,
      question: "How should exceptions and special cases be treated when learning this material?",
      options: [
        "Ignored since they rarely appear in exams",
        "Memorized in isolation without understanding why they exist",
        "Understood as boundary conditions that clarify the general rule",
        "Treated as errors in the theory",
      ],
      correctIndex: 2,
      explanation: "Exceptions and special cases are best understood as boundary conditions that define the limits of general rules. Understanding why they exist deepens your overall comprehension of the material.",
    },
  ]);
});

export default router;