# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StudyMate AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface SignupInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  subject?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  subject?: string;
}

export interface Summary {
  noteId: number;
  summary: string;
  keyPoints: string[];
  generatedAt: string;
}

export interface ExamQuestion {
  id: number;
  question: string;
  answer: string;
  difficulty: string;
}

export interface Flashcard {
  id: number;
  front: string;
  back: string;
  /** @nullable */
  topic?: string | null;
}

export interface QuizQuestion {
  id: number;
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
}

export interface SavedContent {
  id: number;
  noteId: number;
  type: string;
  title: string;
  content: string;
  createdAt: string;
}

export interface SavedContentInput {
  noteId: number;
  type: string;
  title: string;
  content: string;
}

export interface DashboardStats {
  totalNotes: number;
  totalSaved: number;
  summariesGenerated: number;
  flashcardsGenerated: number;
  quizzesGenerated: number;
  recentNotes: Note[];
}