# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudyMate AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: notes
    description: Study notes operations
  - name: ai
    description: AI generation operations
  - name: dashboard
    description: Dashboard operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all notes for the current user
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"

    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note with text content (or from extracted PDF)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /notes/{noteId}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a single note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /notes/{noteId}/summary:
    post:
      operationId: generateSummary
      tags: [ai]
      summary: Generate an AI summary for a note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Summary generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"

  /notes/{noteId}/questions:
    post:
      operationId: generateQuestions
      tags: [ai]
      summary: Generate important exam questions for a note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questions generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamQuestion"

  /notes/{noteId}/flashcards:
    post:
      operationId: generateFlashcards
      tags: [ai]
      summary: Generate flashcards for a note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Flashcards generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"

  /notes/{noteId}/quiz:
    post:
      operationId: generateQuiz
      tags: [ai]
      summary: Generate MCQ quiz for a note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizQuestion"

  /saved:
    get:
      operationId: listSaved
      tags: [notes]
      summary: List all saved generated content for the current user
      responses:
        "200":
          description: Saved content list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedContent"

    post:
      operationId: saveContent
      tags: [notes]
      summary: Save generated AI content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedContentInput"
      responses:
        "201":
          description: Content saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedContent"

  /saved/{savedId}:
    delete:
      operationId: deleteSaved
      tags: [notes]
      summary: Delete saved content
      parameters:
        - name: savedId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics for current user
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required: [id, name, email, createdAt]

    AuthResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]

    SignupInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 6
      required: [name, email, password]

    LoginInput:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required: [email, password]

    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        subject:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, content, createdAt, updatedAt]

    NoteInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        subject:
          type: string
      required: [title, content]

    Summary:
      type: object
      properties:
        noteId:
          type: integer
        summary:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        generatedAt:
          type: string
      required: [noteId, summary, keyPoints, generatedAt]

    ExamQuestion:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        difficulty:
          type: string
      required: [id, question, answer, difficulty]

    Flashcard:
      type: object
      properties:
        id:
          type: integer
        front:
          type: string
        back:
          type: string
        topic:
          type: ["string", "null"]
      required: [id, front, back]

    QuizQuestion:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        explanation:
          type: string
      required: [id, question, options, correctIndex, explanation]

    SavedContent:
      type: object
      properties:
        id:
          type: integer
        noteId:
          type: integer
        type:
          type: string
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required: [id, noteId, type, title, content, createdAt]

    SavedContentInput:
      type: object
      properties:
        noteId:
          type: integer
        type:
          type: string
        title:
          type: string
        content:
          type: string
      required: [noteId, type, title, content]

    DashboardStats:
      type: object
      properties:
        totalNotes:
          type: integer
        totalSaved:
          type: integer
        summariesGenerated:
          type: integer
        flashcardsGenerated:
          type: integer
        quizzesGenerated:
          type: integer
        recentNotes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
      required: [totalNotes, totalSaved, summariesGenerated, flashcardsGenerated, quizzesGenerated, recentNotes]