# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StudyMate AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Note } from './note';

export interface DashboardStats {
  totalNotes: number;
  totalSaved: number;
  summariesGenerated: number;
  flashcardsGenerated: number;
  quizzesGenerated: number;
  recentNotes: Note[];
}