# lib/api-zod/src/generated/types/examQuestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StudyMate AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExamQuestion {
  id: number;
  question: string;
  answer: string;
  difficulty: string;
}