# lib/api-zod/src/generated/types/flashcard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StudyMate AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Flashcard {
  id: number;
  front: string;
  back: string;
  /** @nullable */
  topic?: string | null;
}