# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StudyMate AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  subject?: string | null;
  createdAt: string;
  updatedAt: string;
}