# lib/api-zod/src/generated/types/noteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StudyMate AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  subject?: string;
}