# lib/api-zod/src/generated/types/savedContent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StudyMate AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SavedContent {
  id: number;
  noteId: number;
  type: string;
  title: string;
  content: string;
  createdAt: string;
}