# lib/api-zod/src/generated/types/summary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StudyMate AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Summary {
  noteId: number;
  summary: string;
  keyPoints: string[];
  generatedAt: string;
}